Refactor: update module exports and streamline utility functions
- Updated exports in index.ts to include .js extensions for clarity and added new utility modules.
- Refactored AppDataContext to import computeDataHash from logic instead of utils.
- Simplified authSessionState by re-exporting functions from logic.
- Replaced custom encryption functions in crypto with imports from logic.
- Deleted unused dataHash.ts file to clean up the codebase.
- Updated helpers.tsx to re-export bio dose and raw dose calculation functions from logic.
