+++ packages/core/src/authSessionState.ts
@@ -1,0 +1,7 @@
+let logoutInProgress = false;
+
+export const setLogoutInProgress = (value: boolean) => {
+    logoutInProgress = value;
+};
+
+export const isLogoutInProgress = () => logoutInProgress;
