+++ packages/core/src/index.ts
@@ -1,5 +1,9 @@
-export * from './types';
-export * from './pk';
-export * from './calibration';
-export * from './personalModel';
-export * from './encryption';
+export * from './types.js';
+export * from './pk.js';
+export * from './calibration.js';
+export * from './personalModel.js';
+export * from './encryption.js';
+export * from './dataHash.js';
+export * from './crypto.js';
+export * from './helpers.js';
+export * from './authSessionState.js';
